Correct conditions

+++ .ci/bl-specific.sh
@@ -2,7 +2,7 @@
 
 export USE_EPICS_IOC=0
 
-if [ ! -z "${USE_EPICS_IOC}" ]; then
+if [ ! -z "${USE_EPICS_IOC}" -a "${USE_EPICS_IOC}" -ne 0 ]; then
     docker run -dt --rm --name motorsim --network=host -e "PREFIX=IOC" europeanspallationsource/motorsim
 
     sleep 10
